Refactored Neo-Debugger to fully extract all core functionality into separate assembly to allow for portability to other projects.
