+++ NEO-Debugger-Core/Models/EventArgs.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Neo.Debugger.Models
+namespace Neo.Debugger.Core.Models
 {
     public class CompilerLogEventArgs : EventArgs
     {
