+++ NEO-Debugger-Core/Utils/Compiler.cs
@@ -1,4 +1,4 @@
-﻿using Neo.Debugger.Models;
+﻿using Neo.Debugger.Core.Models;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Neo.Debugger.Utils
+namespace Neo.Debugger.Core.Utils
 {
     public class Compiler
     {
