+++ NEO-Debugger-Core/Utils/Util.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Neo.Debugger.Utils
+namespace Neo.Debugger.Core.Utils
 {
     public static class Util
     {
