+++ NEO-Debugger-Core/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="LunarParser" version="1.0.30" targetFramework="net47" />
+  <package id="Newtonsoft.Json" version="11.0.1" targetFramework="net47" />
+</packages>
