+++ neo-dev-tools.sln
@@ -24,6 +24,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NEO-Compiler-Lib", "NEO-Com
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NEO-IDE", "NEO-Ide\NEO-IDE.csproj", "{8059A07E-609F-4950-872E-F904BCEEB73F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NEO-Debugger-Core", "NEO-Debugger-Core\NEO-Debugger-Core.csproj", "{F034E57E-520B-4572-88F0-76E8D28D4E7B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -66,6 +68,10 @@ Global
 		{8059A07E-609F-4950-872E-F904BCEEB73F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8059A07E-609F-4950-872E-F904BCEEB73F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8059A07E-609F-4950-872E-F904BCEEB73F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F034E57E-520B-4572-88F0-76E8D28D4E7B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F034E57E-520B-4572-88F0-76E8D28D4E7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F034E57E-520B-4572-88F0-76E8D28D4E7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F034E57E-520B-4572-88F0-76E8D28D4E7B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
